Mejora en manejo de fechas a mostrar
